requirements.txt
pytest==7.2.1
pytest-xdist~=3.2.0
coverage~=7.1.0
black~=23.1.0
pytest-timeout~=2.1.0
pytest-env~= 0.8.1
python-dotenv
langchain
langchain-community
langchainhub
langchain-openai 
langchain-chroma
bs4
langchain-core
unstructured
pdf2image
pdfminer.six
langchain_experimental
scikit-learn
ragas==0.1.7
inquirer
llama_index
chromadb
sentence-transformers
llama-index 
llama-index-vector-stores-chroma
llama-index-readers-web
IPython
llama-index-retrievers-bm25
rake_nltk
llama-index-embeddings-langchain
llama-index-vector-stores-faiss
faiss-cpu
llama-index-llms-mistralai
llama-index-embeddings-mistralai
llama-index-embeddings-openai
llama-index-postprocessor-longllmlingua 
llmlingua
llama_index-postprocessor-cohere_rerank
llama_index-postprocessor-jinaai_rerank
llama-index-postprocessor-rankgpt-rerank
llama-index-postprocessor-colbert-rerank
llama-index-postprocessor-rankllm-rerank
llama-index-llms-openai
langchain-huggingface
rank_bm25
ragas
flask
pandas
opencv-python
# unstructured-inference
unstructured
# unstructured[all-docs]
mixpanel
langchain_mistralai
langchain_community
huggingface_hub
datasets
langchain_text_splitters
llama-index-core
requests
markdown
langchain_pinecone
singlestoredb